refactor fastflow loss
