+++ anomalib/models/fastflow/lightning_model.py
@@ -14,7 +14,8 @@ from pytorch_lightning.utilities.cli import MODEL_REGISTRY
 from torch import optim
 
 from anomalib.models.components import AnomalyModule
-from anomalib.models.fastflow.torch_model import FastflowLoss, FastflowModel
+from anomalib.models.fastflow.loss import FastflowLoss
+from anomalib.models.fastflow.torch_model import FastflowModel
 
 logger = logging.getLogger(__name__)
 
@@ -49,7 +50,7 @@ class Fastflow(AnomalyModule):
             conv3x3_only=conv3x3_only,
             hidden_ratio=hidden_ratio,
         )
-        self.loss_func = FastflowLoss()
+        self.loss = FastflowLoss()
 
     def training_step(self, batch, _):  # pylint: disable=arguments-differ
         """Forward-pass input and return the loss.
@@ -62,7 +63,7 @@ class Fastflow(AnomalyModule):
             STEP_OUTPUT: Dictionary containing the loss value.
         """
         hidden_variables, jacobians = self.model(batch["image"])
-        loss = self.loss_func(hidden_variables, jacobians)
+        loss = self.loss(hidden_variables, jacobians)
         return {"loss": loss}
 
     def validation_step(self, batch, _):  # pylint: disable=arguments-differ
