+++ anomalib/models/fastflow/torch_model.py
@@ -86,25 +86,6 @@ def create_fast_flow_block(
     return nodes
 
 
-class FastflowLoss(nn.Module):
-    """FastFlow Loss."""
-
-    def forward(self, hidden_variables: List[Tensor], jacobians: List[Tensor]) -> Tensor:
-        """Calculate the Fastflow loss.
-
-        Args:
-            hidden_variables (List[Tensor]): Hidden variables from the fastflow model. f: X -> Z
-            jacobians (List[Tensor]): Log of the jacobian determinants from the fastflow model.
-
-        Returns:
-            Tensor: _description_
-        """
-        loss = torch.tensor(0.0, device=hidden_variables[0].device)
-        for (hidden_variable, jacobian) in zip(hidden_variables, jacobians):
-            loss += torch.mean(0.5 * torch.sum(hidden_variable**2, dim=(1, 2, 3)) - jacobian)
-        return loss
-
-
 class FastflowModel(nn.Module):
     """FastFlow.
